Description and request for help adding features!!
This "Readme" file is organized into four sections:  "The Input", "The Output", "The Workflow" and "The Missing Features".  They outline how to input metes and bounds, the type of KML that will result, how I usually use the program and finally, what I'd like to add if I had the coding chops, respectively.  Anyone willing to help with one of the four feature requests at the end will be much appreciated and pull requests will be given prompt attention.  Thanks!